refactor: checkbox
