+++ src/packages/checkbox/checkbox.scss
@@ -1,17 +1,5 @@
 @import '../checkboxgroup/checkboxgroup.scss';
 
-.nut-theme-dark {
-  .nut-checkbox {
-    &__label {
-      color: $dark1;
-
-      &--disabled {
-        color: $checkbox-label-disable-color;
-      }
-    }
-  }
-}
-
 .nut-checkbox {
   display: flex;
   align-items: center;
